avoid code duplication of cell vectors hascode and equals
