+++ ydb/core/scheme/scheme_tablecell.cpp
@@ -4,6 +4,8 @@
 #include <ydb/library/actors/core/actorid.h>
 #include <util/string/escape.h>
 
+#include <library/cpp/containers/absl_flat_hash/flat_hash_map.h>
+
 namespace NKikimr {
 
 void TOwnedCellVec::TData::operator delete(void* mem) noexcept {
@@ -62,6 +64,63 @@ TOwnedCellVec::TInit TOwnedCellVec::Allocate(TOwnedCellVec::TCellVec cells) {
         new (mem) TData(),
         size,
     };
+}
+
+struct THashableKey {
+    TConstArrayRef<TCell> Cells;
+
+    template <typename H>
+    friend H AbslHashValue(H h, const THashableKey& key) {
+        h = H::combine(std::move(h), key.Cells.size());
+        for (const TCell& cell : key.Cells) {
+            h = H::combine(std::move(h), cell.IsNull());
+            if (!cell.IsNull()) {
+                h = H::combine(std::move(h), cell.Size());
+                h = H::combine_contiguous(std::move(h), cell.Data(), cell.Size());
+            }
+        }
+        return h;
+    }
+};
+
+size_t TCellVectorsHash::operator()(TConstArrayRef<TCell> key) const {
+    return absl::Hash<THashableKey>()(THashableKey{ key });
+}
+
+bool TCellVectorsEquals::operator() (TConstArrayRef<TCell> a, TConstArrayRef<TCell> b) const {
+    if (a.size() != b.size()) {
+        return false;
+    }
+
+    const TCell* pa = a.data();
+    const TCell* pb = b.data();
+    if (pa == pb) {
+        return true;
+    }
+
+    size_t left = a.size();
+    while (left > 0) {
+        if (pa->IsNull()) {
+            if (!pb->IsNull()) {
+                return false;
+            }
+        } else {
+            if (pb->IsNull()) {
+                return false;
+            }
+            if (pa->Size() != pb->Size()) {
+                return false;
+            }
+            if (pa->Size() > 0 && ::memcmp(pa->Data(), pb->Data(), pa->Size()) != 0) {
+                return false;
+            }
+        }
+        ++pa;
+        ++pb;
+        --left;
+    }
+
+    return true;
 }
 
 namespace {
